+++ src/main/scala/menger/objects/higher_d/Fractal4D.scala
@@ -1,5 +1,5 @@
 package menger.objects.higher_d
 
-trait Fractal4D(val level: Int) extends Mesh4D:
-  override def toString: String = s"${getClass.getSimpleName}(level=$level, ${faces.size} faces)"
+trait Fractal4D(val level: Float) extends Mesh4D:
+  override def toString: String = s"${getClass.getSimpleName}(level=${menger.objects.float2string(level)}, ${faces.size} faces)"